+++ docs/changelog.md
@@ -133,3 +133,5 @@
 - Music added on startup.
 
 - Major shortening of words and a large refractor for stuff to make sense.
+
+- New `math` and `mathround` commands added. Arithmetic only (for now.)